requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pdfminer.six==20240706
google-cloud-storage==2.18.2
openai==1.43.0
